Reverse encryption and compression implemented
Reverse encryption is mostly implemented but decrypting needs work.
Compression implemented and working. On form loading it will create a
Crypty-Chops folder in the roaming folder and try to move the
compression folder into it. If it can't it will just leave it as is and
use it from original folder.